tabix/index/reference_sequence: Fix clippy warnings
This fixes `clippy::iter_kv_map` and `clippy::unnecessary_cast`.

These were added/changed in Rust 1.66.0 / clippy 0.1.66 (69f9c33 2022-12-12).

+++ noodles-tabix/src/index/reference_sequence.rs
@@ -105,11 +105,7 @@ impl ReferenceSequence {
         let (start, end) = resolve_interval(interval)?;
         let region_bins = region_to_bins(start, end);
 
-        let query_bins = self
-            .bins()
-            .iter()
-            .filter(|b| region_bins[b.id() as usize])
-            .collect();
+        let query_bins = self.bins().iter().filter(|b| region_bins[b.id()]).collect();
 
         Ok(query_bins)
     }
